+++ metapackages/auto-instrumentations-node/src/utils.ts
@@ -33,6 +33,7 @@ import { HttpInstrumentation } from '@opentelemetry/instrumentation-http';
 import { IORedisInstrumentation } from '@opentelemetry/instrumentation-ioredis';
 import { KnexInstrumentation } from '@opentelemetry/instrumentation-knex';
 import { KoaInstrumentation } from '@opentelemetry/instrumentation-koa';
+import { LruMemoizerInstrumentation } from '@opentelemetry/instrumentation-lru-memoizer';
 import { MemcachedInstrumentation } from '@opentelemetry/instrumentation-memcached';
 import { MongoDBInstrumentation } from '@opentelemetry/instrumentation-mongodb';
 import { MySQL2Instrumentation } from '@opentelemetry/instrumentation-mysql2';
@@ -41,7 +42,8 @@ import { NestInstrumentation } from '@opentelemetry/instrumentation-nestjs-core'
 import { NetInstrumentation } from '@opentelemetry/instrumentation-net';
 import { PgInstrumentation } from '@opentelemetry/instrumentation-pg';
 import { PinoInstrumentation } from '@opentelemetry/instrumentation-pino';
-import { RedisInstrumentation } from '@opentelemetry/instrumentation-redis';
+import { RedisInstrumentation as RedisInstrumentationV2 } from '@opentelemetry/instrumentation-redis';
+import { RedisInstrumentation as RedisInstrumentationV4 } from '@opentelemetry/instrumentation-redis-4';
 import { RestifyInstrumentation } from '@opentelemetry/instrumentation-restify';
 import { WinstonInstrumentation } from '@opentelemetry/instrumentation-winston';
 
@@ -64,6 +66,7 @@ const InstrumentationMap = {
   '@opentelemetry/instrumentation-ioredis': IORedisInstrumentation,
   '@opentelemetry/instrumentation-knex': KnexInstrumentation,
   '@opentelemetry/instrumentation-koa': KoaInstrumentation,
+  '@opentelemetry/instrumentation-lru-memoizer': LruMemoizerInstrumentation,
   '@opentelemetry/instrumentation-memcached': MemcachedInstrumentation,
   '@opentelemetry/instrumentation-mongodb': MongoDBInstrumentation,
   '@opentelemetry/instrumentation-mysql2': MySQL2Instrumentation,
@@ -72,7 +75,8 @@ const InstrumentationMap = {
   '@opentelemetry/instrumentation-net': NetInstrumentation,
   '@opentelemetry/instrumentation-pg': PgInstrumentation,
   '@opentelemetry/instrumentation-pino': PinoInstrumentation,
-  '@opentelemetry/instrumentation-redis': RedisInstrumentation,
+  '@opentelemetry/instrumentation-redis': RedisInstrumentationV2,
+  '@opentelemetry/instrumentation-redis-4': RedisInstrumentationV4,
   '@opentelemetry/instrumentation-restify': RestifyInstrumentation,
   '@opentelemetry/instrumentation-winston': WinstonInstrumentation,
 };