+++ metapackages/auto-instrumentations-node/test/utils.test.ts
@@ -42,6 +42,7 @@ describe('utils', () => {
         '@opentelemetry/instrumentation-ioredis',
         '@opentelemetry/instrumentation-knex',
         '@opentelemetry/instrumentation-koa',
+        '@opentelemetry/instrumentation-lru-memoizer',
         '@opentelemetry/instrumentation-memcached',
         '@opentelemetry/instrumentation-mongodb',
         '@opentelemetry/instrumentation-mysql2',
@@ -51,10 +52,11 @@ describe('utils', () => {
         '@opentelemetry/instrumentation-pg',
         '@opentelemetry/instrumentation-pino',
         '@opentelemetry/instrumentation-redis',
+        '@opentelemetry/instrumentation-redis-4',
         '@opentelemetry/instrumentation-restify',
         '@opentelemetry/instrumentation-winston',
       ];
-      assert.strictEqual(instrumentations.length, 28);
+      assert.strictEqual(instrumentations.length, 30);
       for (let i = 0, j = instrumentations.length; i < j; i++) {
         assert.strictEqual(
           instrumentations[i].instrumentationName,