+++ plugins/node/opentelemetry-instrumentation-ioredis/test/ioredis.test.ts
@@ -43,6 +43,7 @@ import {
   DbSystemValues,
   SemanticAttributes,
 } from '@opentelemetry/semantic-conventions';
+import { defaultDbStatementSerializer } from '../src/utils';
 
 const memoryExporter = new InMemorySpanExporter();
 
@@ -187,12 +188,14 @@ describe('ioredis', () => {
       description: string;
       name: string;
       args: Array<string>;
+      serializedArgs: Array<string>;
       method: (cb: ioredisTypes.CallbackFunction<unknown>) => unknown;
     }> = [
       {
         description: 'insert',
         name: 'hset',
         args: [hashKeyName, 'testField', 'testValue'],
+        serializedArgs: [hashKeyName, 'testField', '[1 other arguments]'],
         method: (cb: ioredisTypes.CallbackFunction<number>) =>
           client.hset(hashKeyName, 'testField', 'testValue', cb),
       },
@@ -200,6 +203,7 @@ describe('ioredis', () => {
         description: 'get',
         name: 'get',
         args: [testKeyName],
+        serializedArgs: [testKeyName],
         method: (cb: ioredisTypes.CallbackFunction<string | null>) =>
           client.get(testKeyName, cb),
       },
@@ -243,7 +247,7 @@ describe('ioredis', () => {
             ...DEFAULT_ATTRIBUTES,
             [SemanticAttributes.DB_STATEMENT]: `${
               command.name
-            } ${command.args.join(' ')}`,
+            } ${command.serializedArgs.join(' ')}`,
           };
           const span = provider
             .getTracer('ioredis-test')
@@ -273,7 +277,7 @@ describe('ioredis', () => {
       it('should create a child span for hset promise', async () => {
         const attributes = {
           ...DEFAULT_ATTRIBUTES,
-          [SemanticAttributes.DB_STATEMENT]: `hset ${hashKeyName} random random`,
+          [SemanticAttributes.DB_STATEMENT]: `hset ${hashKeyName} random [1 other arguments]`,
         };
         const span = provider.getTracer('ioredis-test').startSpan('test span');
         await context.with(trace.setSpan(context.active(), span), async () => {
@@ -466,7 +470,7 @@ describe('ioredis', () => {
       it('should create a child span for pipeline', done => {
         const attributes = {
           ...DEFAULT_ATTRIBUTES,
-          [SemanticAttributes.DB_STATEMENT]: 'set foo bar',
+          [SemanticAttributes.DB_STATEMENT]: 'set foo [1 other arguments]',
         };
 
         const span = provider.getTracer('ioredis-test').startSpan('test span');
@@ -636,16 +640,24 @@ describe('ioredis', () => {
         };
         instrumentation.setConfig(config);
       });
-      it('should not create child span', async () => {
+      it('should not create child span for query', async () => {
         await client.set(testKeyName, 'data');
         const result = await client.del(testKeyName);
         assert.strictEqual(result, 1);
         assert.strictEqual(memoryExporter.getFinishedSpans().length, 0);
       });
+
+      it('should not create child span for connect', async () => {
+        const lazyClient = new ioredis(URL, { lazyConnect: true });
+        await lazyClient.connect();
+        const spans = memoryExporter.getFinishedSpans();
+        await lazyClient.quit();
+        assert.strictEqual(spans.length, 0);
+      });
     });
 
     describe('Instrumentation with requireParentSpan', () => {
-      it('should instrument with requireParentSpan equal false', async () => {
+      it('should instrument queries with requireParentSpan equal false', async () => {
         const config: IORedisInstrumentationConfig = {
           requireParentSpan: false,
         };
@@ -654,23 +666,48 @@ describe('ioredis', () => {
         await client.set(testKeyName, 'data');
         const result = await client.del(testKeyName);
         assert.strictEqual(result, 1);
-
         const endedSpans = memoryExporter.getFinishedSpans();
-        assert.strictEqual(endedSpans.length, 2);
 
+        assert.strictEqual(endedSpans.length, 2);
         testUtils.assertSpan(
           endedSpans[0],
           SpanKind.CLIENT,
           {
             ...DEFAULT_ATTRIBUTES,
-            [SemanticAttributes.DB_STATEMENT]: `set ${testKeyName} data`,
+            [SemanticAttributes.DB_STATEMENT]: `set ${testKeyName} [1 other arguments]`,
           },
           [],
           unsetStatus
         );
       });
 
-      it('should not instrument with requireParentSpan equal true', async () => {
+      it('should instrument connect with requireParentSpan equal false', async () => {
+        const config: IORedisInstrumentationConfig = {
+          requireParentSpan: false,
+        };
+        instrumentation.setConfig(config);
+
+        const lazyClient = new ioredis(URL, { lazyConnect: true });
+        await lazyClient.connect();
+        const endedSpans = memoryExporter.getFinishedSpans();
+        assert.strictEqual(endedSpans.length, 2);
+        assert.strictEqual(endedSpans[0].name, 'connect');
+        assert.strictEqual(endedSpans[1].name, 'info');
+
+        await lazyClient.quit();
+        testUtils.assertSpan(
+          endedSpans[0],
+          SpanKind.CLIENT,
+          {
+            ...DEFAULT_ATTRIBUTES,
+            [SemanticAttributes.DB_STATEMENT]: 'connect',
+          },
+          [],
+          unsetStatus
+        );
+      });
+
+      it('should not instrument queries with requireParentSpan equal true', async () => {
         const config: IORedisInstrumentationConfig = {
           requireParentSpan: true,
         };
@@ -681,6 +718,20 @@ describe('ioredis', () => {
         assert.strictEqual(result, 1);
 
         assert.strictEqual(memoryExporter.getFinishedSpans().length, 0);
+      });
+
+      it('should not instrument connect with requireParentSpan equal true', async () => {
+        const config: IORedisInstrumentationConfig = {
+          requireParentSpan: true,
+        };
+        instrumentation.setConfig(config);
+
+        const lazyClient = new ioredis(URL, { lazyConnect: true });
+        await lazyClient.connect();
+        const endedSpans = memoryExporter.getFinishedSpans();
+        assert.strictEqual(endedSpans.length, 0);
+
+        await lazyClient.quit();
       });
     });
 
@@ -948,6 +999,43 @@ describe('ioredis', () => {
             });
           });
         });
+      });
+    });
+  });
+
+  describe('#defaultDbStatementSerializer()', () => {
+    [
+      {
+        cmdName: 'UNKNOWN',
+        cmdArgs: ['something'],
+        expected: 'UNKNOWN [1 other arguments]',
+      },
+      {
+        cmdName: 'ECHO',
+        cmdArgs: ['echo'],
+        expected: 'ECHO [1 other arguments]',
+      },
+      {
+        cmdName: 'LPUSH',
+        cmdArgs: ['list', 'value'],
+        expected: 'LPUSH list [1 other arguments]',
+      },
+      {
+        cmdName: 'HSET',
+        cmdArgs: ['hash', 'field', 'value'],
+        expected: 'HSET hash field [1 other arguments]',
+      },
+      {
+        cmdName: 'INCRBY',
+        cmdArgs: ['key', 5],
+        expected: 'INCRBY key 5',
+      },
+    ].forEach(({ cmdName, cmdArgs, expected }) => {
+      it(`should serialize the correct number of arguments for ${cmdName}`, () => {
+        assert.strictEqual(
+          defaultDbStatementSerializer(cmdName, cmdArgs),
+          expected
+        );
       });
     });
   });