+++ plugins/node/opentelemetry-instrumentation-mysql2/src/utils.ts
@@ -16,7 +16,23 @@
 
 import { SpanAttributes } from '@opentelemetry/api';
 import { SemanticAttributes } from '@opentelemetry/semantic-conventions';
-import type { Query, QueryOptions } from 'mysql2';
+
+/*
+  Following types declare an expectation on mysql2 types and define a subset we
+  use in the instrumentation of the types actually defined in mysql2 pacakge
+
+  We need to import them here so that the installing party of the instrumentation
+  doesn't have to absolutely install the mysql2 package as well - specially
+  important for auto-loaders and meta-pacakges.
+*/
+interface QueryOptions {
+  sql: string;
+  values?: any | any[] | { [param: string]: any };
+}
+
+interface Query {
+  sql: string;
+}
 
 interface Config {
   host?: string;