+++ plugins/node/opentelemetry-instrumentation-router/src/instrumentation.ts
@@ -33,9 +33,7 @@ import * as utils from './utils';
 import AttributeNames from './enums/AttributeNames';
 import LayerType from './enums/LayerType';
 
-export default class RouterInstrumentation extends InstrumentationBase<
-  typeof Router
-> {
+export default class RouterInstrumentation extends InstrumentationBase<any> {
   constructor() {
     super(`@opentelemetry/instrumentation-${constants.MODULE_NAME}`, VERSION);
   }
@@ -43,7 +41,7 @@ export default class RouterInstrumentation extends InstrumentationBase<
   private _moduleVersion?: string;
 
   init() {
-    const module = new InstrumentationNodeModuleDefinition<typeof Router>(
+    const module = new InstrumentationNodeModuleDefinition<any>(
       constants.MODULE_NAME,
       constants.SUPPORTED_VERSIONS,
       (moduleExports, moduleVersion) => {
@@ -170,7 +168,7 @@ export default class RouterInstrumentation extends InstrumentationBase<
     const type = layer.method
       ? LayerType.REQUEST_HANDLER
       : LayerType.MIDDLEWARE;
-    const route = utils.getRoute(req);
+    const route = req.baseUrl + (req.route?.path ?? '') || '/';
     const spanName =
       type === LayerType.REQUEST_HANDLER
         ? `request handler - ${route}`