+++ propagators/opentelemetry-propagator-ot-trace/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [0.26.1](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/propagator-ot-trace-v0.26.0...propagator-ot-trace-v0.26.1) (2022-08-09)
+
+
+### Bug Fixes
+
+* **propagator-ot-trace:** read sampled flag correctly from span context ([#1077](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1077)) ([69740ab](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/69740aba848486908e924376e3ca093ab88720b6))
+
+## [0.26.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/propagator-ot-trace-v0.25.1...propagator-ot-trace-v0.26.0) (2022-05-14)
+
+
+### Features
+
+* add supported node versions for all packages ([#973](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/973)) ([baaacbd](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/baaacbdd35ca4baab0afae64647aa8c0380ee4b7))
+* remove colors dependency ([#943](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/943)) ([b21b96c](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/b21b96c1a3a4f871370f970d6b2825f00e1fe595)), closes [#826](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/826)
+* update webpack outside of examples ([#963](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/963)) ([9a58648](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/9a586480ed6a7677fb1283a61d05540345c52617))
+* use Otel SDK 1.2/0.28 ([#984](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/984)) ([098c2ed](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/098c2ed6f9c5ab7bd865685018c0777245aab3b7))
+
 ### [0.25.1](https://www.github.com/open-telemetry/opentelemetry-js-contrib/compare/propagator-ot-trace-v0.25.0...propagator-ot-trace-v0.25.1) (2022-01-24)